Run ExceptionsManager unit tests in both __DEV__ and prod mode
Summary:
Changelog:
[General][Fixed] - Run ExceptionsManager unit tests in both __DEV__ and prod mode

This suite of unittests use a `__DEV__` check in various places to check that the correct data is sent to either LogBox (in __DEV__) or the Native ExceptionsManager (in PROD). However, we're not actually running both modes, so half of the assert branches in these tests aren't ever utilized.

This diff uses the same approach as ReactNativeVersionCheck-test.js to execute this test suite twice: Once in DEV and once in PROD mode.

Reviewed By: motiz88

Differential Revision: D36099190

fbshipit-source-id: 40b8ea26f1d9e093202f3c3f3b55111110a8d64c